requirements.txt
google-generativeai
pillow
python-dotenv
streamlit
pytesseract
